Fix compiler warnings/errors
In a few cases, we passed a function result of type Datum to a
type-specific PG_RETURN_* macro.  This used to work because underneath
everything was cast to the right type.  Since PostgreSQL 16, some of
the underlying macros are now inline functions and enforce correct
types.  So these now caused compiler warnings, apparently only
warnings on gcc, but I suppose everything still worked, at least on
64-bit platforms.  Under clang, these are errors, which causes the
bitcode builds to fail.

The fix is to use PG_RETURN_DATUM instead.